move authentication to public
