+++ src/test/java/de/tum/in/www1/artemis/AbstractSpringIntegrationSaml2Test.java
@@ -25,10 +25,10 @@ import de.tum.in.www1.artemis.service.connectors.SAML2Service;
 import de.tum.in.www1.artemis.service.connectors.jenkins.JenkinsProgrammingLanguageFeatureService;
 import de.tum.in.www1.artemis.util.DatabaseUtilService;
 import de.tum.in.www1.artemis.util.RequestUtilService;
-import de.tum.in.www1.artemis.web.rest.UserJWTResource;
+import de.tum.in.www1.artemis.web.rest.publicc.UserJwtResource;
 
 /**
- * Test base for {@link UserJWTResource#authorizeSAML2(String)} and {@link SAML2Service}.
+ * Test base for {@link UserJwtResource#authorizeSAML2(String)} and {@link SAML2Service}.
  *
  * @author Dominik Fuchss
  */
