+++ src/test/java/de/tum/in/www1/artemis/authentication/InternalAuthenticationIntegrationTest.java
@@ -34,8 +34,8 @@ import de.tum.in.www1.artemis.security.Role;
 import de.tum.in.www1.artemis.security.jwt.TokenProvider;
 import de.tum.in.www1.artemis.service.user.PasswordService;
 import de.tum.in.www1.artemis.util.ModelFactory;
-import de.tum.in.www1.artemis.web.rest.UserJWTResource;
 import de.tum.in.www1.artemis.web.rest.dto.LtiLaunchRequestDTO;
+import de.tum.in.www1.artemis.web.rest.publicc.UserJwtResource;
 import de.tum.in.www1.artemis.web.rest.vm.LoginVM;
 import de.tum.in.www1.artemis.web.rest.vm.ManagedUserVM;
 
@@ -257,7 +257,7 @@ class InternalAuthenticationIntegrationTest extends AbstractSpringIntegrationJen
         HttpHeaders httpHeaders = new HttpHeaders();
         httpHeaders.add("User-Agent", "Mozilla/5.0 (X11; Linux x86_64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/73.0.3683.103 Safari/537.36");
 
-        UserJWTResource.JWTToken jwtToken = request.postWithResponseBody("/api/authenticate", loginVM, UserJWTResource.JWTToken.class, HttpStatus.OK, httpHeaders);
+        UserJwtResource.JWTToken jwtToken = request.postWithResponseBody("/api/public/authenticate", loginVM, UserJwtResource.JWTToken.class, HttpStatus.OK, httpHeaders);
         assertThat(jwtToken.getIdToken()).as("JWT token is present").isNotNull();
         assertThat(this.tokenProvider.validateTokenForAuthority(jwtToken.getIdToken())).as("JWT token is valid").isTrue();
     }
